미션 6 정답지로 덮어쓰기

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicUserService.java
@@ -1,12 +1,13 @@
 package com.sprint.mission.discodeit.service.basic;
 
-import com.sprint.mission.discodeit.dto.data.UserDTO;
-import com.sprint.mission.discodeit.dto.user.BinaryContentCreateRequest;
-import com.sprint.mission.discodeit.dto.user.UserCreateRequest;
-import com.sprint.mission.discodeit.dto.user.UserUpdateRequest;
+import com.sprint.mission.discodeit.dto.data.UserDto;
+import com.sprint.mission.discodeit.dto.request.BinaryContentCreateRequest;
+import com.sprint.mission.discodeit.dto.request.UserCreateRequest;
+import com.sprint.mission.discodeit.dto.request.UserUpdateRequest;
 import com.sprint.mission.discodeit.entity.BinaryContent;
 import com.sprint.mission.discodeit.entity.User;
 import com.sprint.mission.discodeit.entity.UserStatus;
+import com.sprint.mission.discodeit.mapper.UserMapper;
 import com.sprint.mission.discodeit.repository.BinaryContentRepository;
 import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.repository.UserStatusRepository;
@@ -21,21 +22,20 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-@Transactional
 @RequiredArgsConstructor
 @Service
 public class BasicUserService implements UserService {
 
   private final UserRepository userRepository;
-  private final BinaryContentRepository binaryContentRepository;
   private final UserStatusRepository userStatusRepository;
-  private final BinaryContentStorage storage;
+  private final UserMapper userMapper;
+  private final BinaryContentRepository binaryContentRepository;
+  private final BinaryContentStorage binaryContentStorage;
 
+  @Transactional
   @Override
-  public User create(
-      UserCreateRequest userCreateRequest,
-      Optional<BinaryContentCreateRequest> optionalProfileCreateRequest
-  ) {
+  public UserDto create(UserCreateRequest userCreateRequest,
+      Optional<BinaryContentCreateRequest> optionalProfileCreateRequest) {
     String username = userCreateRequest.username();
     String email = userCreateRequest.email();
 
@@ -46,62 +46,52 @@ public class BasicUserService implements UserService {
       throw new IllegalArgumentException("User with username " + username + " already exists");
     }
 
-    BinaryContent nullableProfileId = optionalProfileCreateRequest.map(
-        profileRequest -> {
+    BinaryContent nullableProfile = optionalProfileCreateRequest
+        .map(profileRequest -> {
           String fileName = profileRequest.fileName();
           String contentType = profileRequest.contentType();
           byte[] bytes = profileRequest.bytes();
-
-          BinaryContent binaryContent = binaryContentRepository.save(
-              new BinaryContent(fileName, (long) bytes.length, contentType));
-
-          storage.put(binaryContent.getId(), bytes);
+          BinaryContent binaryContent = new BinaryContent(fileName, (long) bytes.length,
+              contentType);
+          binaryContentRepository.save(binaryContent);
+          binaryContentStorage.put(binaryContent.getId(), bytes);
           return binaryContent;
-
-        }).orElse(null);
-
+        })
+        .orElse(null);
     String password = userCreateRequest.password();
+
+    User user = new User(username, email, password, nullableProfile);
     Instant now = Instant.now();
+    UserStatus userStatus = new UserStatus(user, now);
 
-//    User createdUser = userRepository.save(new User(username, email, password, nullableProfileId));
-//    userStatusRepository.save(new UserStatus(createdUser, now));
-
-    User createdUser = new User(username, email, password, nullableProfileId);
-    UserStatus status = new UserStatus(createdUser, now);
-    createdUser.setUserStatus(status);   // ★ 이 줄이 핵심
-
-    userRepository.save(createdUser);     // cascade로 함께 저장
-
-
-    return createdUser;
+    userRepository.save(user);
+    return userMapper.toDto(user);
   }
 
   @Override
-  public UserDTO find(UUID userId) {
-    return userRepository.findById(userId).map(this::toDto)
+  public UserDto find(UUID userId) {
+    return userRepository.findById(userId)
+        .map(userMapper::toDto)
         .orElseThrow(() -> new NoSuchElementException("User with id " + userId + " not found"));
   }
 
   @Override
-  public UserDTO findByUsername(String username) {
-    return userRepository.findByUsername(username).map(this::toDto)
-        .orElseThrow(() -> new NoSuchElementException("User with id " + username + " not found"));
+  public List<UserDto> findAll() {
+    return userRepository.findAllWithProfileAndStatus()
+        .stream()
+        .map(userMapper::toDto)
+        .toList();
   }
 
+  @Transactional
   @Override
-  public List<UserDTO> findAll() {
-    return userRepository.findAll().stream().map(this::toDto).toList();
-  }
-
-  @Override
-  public User update(UUID userId, UserUpdateRequest userUpdateRequest,
+  public UserDto update(UUID userId, UserUpdateRequest userUpdateRequest,
       Optional<BinaryContentCreateRequest> optionalProfileCreateRequest) {
     User user = userRepository.findById(userId)
         .orElseThrow(() -> new NoSuchElementException("User with id " + userId + " not found"));
 
     String newUsername = userUpdateRequest.newUsername();
     String newEmail = userUpdateRequest.newEmail();
-
     if (userRepository.existsByEmail(newEmail)) {
       throw new IllegalArgumentException("User with email " + newEmail + " already exists");
     }
@@ -109,45 +99,33 @@ public class BasicUserService implements UserService {
       throw new IllegalArgumentException("User with username " + newUsername + " already exists");
     }
 
-    BinaryContent nullableProfile = optionalProfileCreateRequest.map(profileRequest -> {
-      Optional.ofNullable(user.getProfile().getId())
-          .ifPresent(binaryContentRepository::deleteById);
+    BinaryContent nullableProfile = optionalProfileCreateRequest
+        .map(profileRequest -> {
 
-      String fileName = profileRequest.fileName();
-      String contentType = profileRequest.contentType();
-
-      byte[] bytes = profileRequest.bytes();
-      BinaryContent binaryContent =
-          binaryContentRepository.save(new BinaryContent(
-              fileName, (long) bytes.length, contentType
-          ));
-
-      storage.put(binaryContent.getId(), bytes);
-      return binaryContent;
-    }).orElse(null);
+          String fileName = profileRequest.fileName();
+          String contentType = profileRequest.contentType();
+          byte[] bytes = profileRequest.bytes();
+          BinaryContent binaryContent = new BinaryContent(fileName, (long) bytes.length,
+              contentType);
+          binaryContentRepository.save(binaryContent);
+          binaryContentStorage.put(binaryContent.getId(), bytes);
+          return binaryContent;
+        })
+        .orElse(null);
 
     String newPassword = userUpdateRequest.newPassword();
     user.update(newUsername, newEmail, newPassword, nullableProfile);
 
-    return userRepository.save(user);
+    return userMapper.toDto(user);
   }
 
+  @Transactional
   @Override
   public void delete(UUID userId) {
-    User user = userRepository.findById(userId)
-        .orElseThrow(() -> new NoSuchElementException("User with id " + userId + " not found"));
-
-    Optional.ofNullable(user.getProfile().getId()).ifPresent(binaryContentRepository::deleteById);
-    userStatusRepository.deleteByUserId(userId);
+    if (userRepository.existsById(userId)) {
+      throw new NoSuchElementException("User with id " + userId + " not found");
+    }
 
     userRepository.deleteById(userId);
-  }
-
-  private UserDTO toDto(User user) {
-    Boolean online = userStatusRepository.findByUserId(user.getId()).map(UserStatus::isOnline)
-        .orElse(null);
-
-    return new UserDTO(user.getId(), user.getUsername(),
-        user.getEmail(), user.getProfile().getId(), online);
   }
 }
